refactor(handoff): isolate transaction boundaries
