+++ daemon/handoff.go
@@ -1,14 +1,12 @@
 package daemon
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"strings"
 
 	"github.com/sachiniyer/agent-factory/log"
 	"github.com/sachiniyer/agent-factory/session"
-	"github.com/sachiniyer/agent-factory/task"
 )
 
 // HandoffSessionRequest asks the daemon to continue a session under a different
@@ -234,62 +232,11 @@ func (m *Manager) HandoffSession(req HandoffSessionRequest) (HandoffSessionRespo
 		log.WarningLog.Printf("handoff %q: failed to persist the post-swap checkpoint before mission delivery: %v", req.Title, err)
 	}
 
-	settle := func(event session.TransitionEvent, clearPending bool) error {
-		if err := instance.Transition(event); err != nil {
-			return err
-		}
-		if clearPending && !instance.ClearPendingHandoffMission(mission) {
-			return fmt.Errorf("pending handoff mission changed before settlement")
-		}
-		// Persist only settled outcomes. Disk persistence strips transient ops, so
-		// writing while OpReplacing is raised would manufacture an unfenced state
-		// no in-memory reader was ever allowed to observe.
-		m.persistInstance(repoID, instance)
-		m.captureAgentConversationAsync(repoID, key, instance, conversationCapture)
-		return nil
-	}
-	if serr := task.WaitForReadyAndSendPrompt(context.Background(), instance, mission); serr != nil {
-		var limitErr *task.LimitReachedError
-		if errors.As(serr, &limitErr) {
-			// The mission never reached the composer. Store the rendered takeover
-			// context — not the old create prompt or bare override — because the
-			// normal limit-resume path replays exactly Instance.Prompt.
-			instance.SetPrompt(mission)
-			if terr := settle(session.ParkHandoff(limitErr.ResetAt), true); terr != nil {
-				return HandoffSessionResponse{}, fmt.Errorf(
-					"handed %q off to %s, which hit a usage limit before its mission could be delivered, and failed to park the handoff: %w",
-					req.Title, target, errors.Join(serr, terr))
-			}
-			return HandoffSessionResponse{}, fmt.Errorf(
-				"handed %q off to %s, which hit a usage limit before its mission could be delivered; "+
-					"the rendered mission is saved and will be sent by the normal limit-resume path: %w",
-				req.Title, target, serr)
-		}
-		if !errors.Is(serr, task.ErrPromptDelivery) {
-			// Readiness/trust failures do not prove a usable incoming runtime. Retain
-			// the record and pending mission, but make it inert: claiming Running here
-			// turns a startup death into a fake delivery failure and invites commands
-			// into an unconfirmed binding.
-			instance.MarkStartupStateUnknown()
-			m.persistInstance(repoID, instance)
-			return HandoffSessionResponse{}, fmt.Errorf(
-				"handed %q off to %s, but the incoming agent never reached a confirmed ready state (%w); "+
-					"the session was retained as startup-unknown with its mission pending for inspection",
-				req.Title, target, serr)
-		}
-		// Readiness proved this is the incoming runtime, so conversation capture
-		// is safe even though its mission remains behind the durable retry fence.
-		// The later recovery pass no longer has the command-specific capture plan.
-		m.captureAgentConversationAsync(repoID, key, instance, conversationCapture)
-		return HandoffSessionResponse{}, fmt.Errorf(
-			"handed %q off to %s, but its mission brief could not be delivered (%w); "+
-				"the exact mission remains pending behind the replacement fence for a readiness-based retry",
-			req.Title, target, serr)
-	}
-	if err := settle(session.CommitHandoff(), true); err != nil {
-		return HandoffSessionResponse{}, fmt.Errorf(
-			"handed %q off to %s and delivered its mission, but could not settle the replacement fence: %w",
-			req.Title, target, err)
+	if err := m.deliverHandoffMission(handoffDelivery{
+		repoID: repoID, key: key, title: req.Title, target: target, mission: mission,
+		instance: instance, conversationCapture: conversationCapture,
+	}); err != nil {
+		return HandoffSessionResponse{}, err
 	}
 	log.InfoLog.Printf("handoff: session %q swapped %s → %s at %s", instance.Title, outgoing, target, shortSHA(headSHA))
